rms.h: calculate RMS properly
All samples should be squared as is, instead of taking average between
the channels.  (This can be slightly slower; on the other hand it can be
vectorized; I have not looked into this yet.)  The result is no longer
adjusted for the number of channels.  It does not make much difference,
because in waveformgen.h, levels are calibrated against q95 and q999.